fix(ci): python-full-coverage — split oversized test_config_part_001 to honor 1500-line guardrail
test_config_part_001.py had grown to 1513 lines, tripping the
test_first_party_code_files_stay_under_line_limit maintainability
guardrail (MAX_FIRST_PARTY_FILE_LINES = 1500). Moved the trailing block
of service-settings tests (agent watchdog, planning iterations, disk
threshold, log tail, work-dir resolution, cleanup policy, network
posture, node id) into a new test_config_part_003.py, mirroring the
existing parts split, and dropped the now-unused imports from part_001.
No tests were removed, skipped, or weakened.
